requirements.txt
flask>=3.0.0
python-dotenv>=1.0.0
yt-dlp>=2024.1.0

# Topic Modeling
scikit-learn>=1.3.0
gensim>=4.3.0
bertopic>=0.16.0
sentence-transformers>=2.2.0
transformers>=4.38.0
torch>=2.0.0

# NLP & Pre-traitement
spacy>=3.7.0
nltk>=3.8.0

# Reduction dimensionnelle
umap-learn>=0.5.0

# Visualisation (Plotly est charge via CDN cote client)
# plotly>=5.18.0
